\name{FormatData}
\alias{FormatData}
\title{Formats the data for NBDA}
\usage{
FormatData(a, b, c, d, e, f, g,h,networkdata)
}
\arguments{
  \item{a}{diffusion times}

  \item{b}{social network}

  \item{c}{unique id for individual at that particular event}

  \item{d}{group number}

  \item{e}{diffusion number}

  \item{f}{event number}

  \item{g}{ alternative social network}
  
  \item{h}{spatial covariate}
  \item{networkdata}{list of social networks}
}
\description{
Formats the data for NBDA
}

