\name{Ids}
\alias{Ids}
\docType{data}
\title{
This dataset contains the unique id for each individual in each group.
This is part of the data example to demonstrate modelling NBDA with random effects.
}

\description{
The ids are the unique id for each individual that was observed to first solve a task
or demonstrate a new behaviour.
}

\usage{data(Ids)}



