\name{nullmcmc}
\alias{nullmcmc}

\title{
The spatial NBDA null model is considered for this analysis.
The baseline rate parameter is estimated.
}
\description{
The model considered contains only one parameter: the baseline rate parameter, and represents the rate at 
which the behaviour under consideration is acquired.

}
\usage{
nullmcmc(formatteddata, its, pilot_tuner, start)
}

\arguments{
  \item{formatteddata}{
Formatted data using the FormatData function.
}
  \item{its}{
Number of iterations.
}
  \item{pilot_tuner}{
Tuning parameter
}

 \item{start}{
Start value.
}
 
}
 

