\name{papernests}
\alias{papernests}
\docType{data}
\title{
papernests
}
\description{
This dataset provides the unique id of each individual nest in the order of the first observation of the behaviour under consideration.  This is for the spatial NBDA example with an environmental covariate..
}
\usage{data(papernests)}
\format{
  A data frame with 70 observations.
  \describe{
    \item{\code{Nests}}{a numeric vector}
  }
}


\examples{
data(papernests)
}
\keyword{datasets}
