\name{papertimes}
\alias{papertimes}
\docType{data}
\title{
papertimes
}
\description{
This dataset provides the diffusion times for the dataset used for spatial NBDA example with an
environmental covariate..  
The values represent the times at which the behaviour under consideration was first observed at a given nest.
}
\usage{data(papertimes)}
\format{
  A data frame with 70 observations.
  \describe{
    \item{\code{Times}}{a numeric vector}
  }
}

\examples{
data(papertimes)
## maybe str(papertimes) ; plot(papertimes) ...
}
\keyword{datasets}
