\name{Times}
\alias{Times}
\docType{data}
\title{
This dataset contains the diffusion times using in the 
This is part of the data example to demonstrate modelling NBDA with random effects.
}

\description{
The times are the times for which each individual was first observed to  solve a task
or demonstrate a new behaviour.
}

\usage{data(Times)}



